fixed issue when container is already running

+++ ng.sh
@@ -29,4 +29,10 @@ else
     CMD="ng $@"
 fi
 
-docker run --rm --name $IMAGE_NAME -it -p 4200:4200 -v $PWD:/app:rw -v ~/.gitconfig:/home/angular/.gitconfig:ro -h $IMAGE_NAME $IMAGE_NAME $CMD
+docker ps | grep $IMAGE_NAME > /dev/null
+if [ "$?" -eq 0 ]; then
+    docker exec -it $IMAGE_NAME $CMD 
+else
+    docker run --rm --name $IMAGE_NAME -it -p 4200:4200 -v $PWD:/app:rw -v ~/.gitconfig:/home/angular/.gitconfig:ro -h $IMAGE_NAME $IMAGE_NAME $CMD
+fi
+
